Move preserving history
